Determine cell proof support by checking L1 cfg
This change modifies --node.batch-poster.data-poster.enable-cell-proofs
to take a string which is either "" or "auto" indicating to auto detect
the cell proof support based on L1 Osaka fork status, or "force-enable"
or "force-disable" which overrides the auto detected status, which may
be useful for testing or emergencies.

NOTE: Our currently imported go-ethereum version does not have OsakaTime
yet in its chain config for Mainnet, although it has it for Sepolia. So
we will need to update our imported go-ethereum version before the
mainnet fork activation or plan to set the configuration manually.