Hide blob-related DataPoster config in some contexts (#3810)
* Hide blob-related DataPoster config in some contexts

The DataPoster component is reused for the BatchPoster and the Staker,
but not all of its configuration options are applicable in both
contexts.

The Staker should never be configured to post its assertions as blobs,
so this change prevents that, and also removes blob tuning related
options. On a properly configured parent chain the Staker probably never
would have decided to post a blob even if configured to do so since they
are small and it would waste most of the space in the blob. On the
outside chance that a blob was posted, Nitro doesn't have code for
reading it. This change prevents operator confusion, increases safety and
also means the DataPoster in the Staker can skip any 4844 related checks.

Removed Staker config:
```
--node.staker.data-poster.post-4844-blobs
--node.staker.data-poster.blob-tx-replacement-times
--node.staker.data-poster.min-blob-tx-tip-cap-gwei
--node.staker.data-poster.max-blob-tx-tip-cap-gwei
--node.staker.data-poster.enable-cell-proofs
```

The BatchPoster's DataPoster also used to have a redundant configuration
that has been removed since the BatchPoster's configuration overrides it
(see dataPosterConfigFetcher in batch_poster.go). This configuration
option was never actually used by the code so it's better to remove it.

Removed BatchPoster config:
```
--node.batch-poster.data-poster.post-4844-blobs
```

The existing config option that should be used is:
```
--node.batch-poster.post-4844-blobs
```

NOTE: Nitro will fail to start if the removed flags are used so
operators may need to update their configuration files if they were
using the old incorrect flags.

* Remove dead code

* Set empty defaults for blob fields on validator default config