Correct the order TransactionStreamer and PopulateFeedBacklog (#3812)
The TransactionStreamer and InboxReader are now needed by PopulateFeedBacklog as
of c6bd24a becuase the GetMessage implementation in the transaction streamer
needs the InboxReader to have been started so it can get the blobs associated
with a message.

This is problematic becuase during initialization of the node
PopulateFeedBacklog uses the transaction streamer to get messages which were
stored in the database when there is a backlog.

Fixes: NIT-3997