Simplify alt DA configuration
Remove --node.da.mode command line option. .Enable on the desired DA
option is used instead.  This makes it possible to .Enable both Anytrust
and External DA.

Remove --node.da.referenceda entirely. There was test code that was
starting an embedded referenceda provider, but this is not a
configuration that should be used in production so the test was modified
to start an external referenceda provider.