Fix Anytrust test flakiness around fetching keyset
KeysetFetcher.GetKeysetByHash's logic for finding the SetValidKeyset
event where the keyset was created was written assuming that the
eth_getLogs From and To range was exclusive on the To side, but it's
actually inclusive. Even if tests waited for the keyset creation tx to
be included in a block, the KeysetFetcher would be trying to query the
next block too, which may not exist yet.

This change simply treats the block range as inclusive, and also adds an
additional retry loop to the test function authorizeDASKeyset that
retries until the log for the keyset creation is actually queryable.