Fix Anytrust DisableSignatureChecking=true
DisableSignatureChecking=true is used in some tests and for manual
testing.

Previously on the DAS server side the SignatureVerifier was being
created with a nil seqInboxCaller in the case where signature checking
was disabled. This caused "no signature verification method configured"
errors. Now the SignatureVerifier is not created and the RPC server uses
the NoopPayloadVerifier for the streaming protocol when signature
checking is disabled.

On the client side we add a check that there is a signer if the
streaming protocol is to be used, since the signer uses the signatures
for replay protection.