Explicit fallback for DA provider failures
This commit changes the fallback behavior between DA providers in the
BatchPoster to require an explicit signal to fall back via
ErrFallbackRequested. In previous commits we implemented a resizing
fallback mechanism so that if a batch is too big for EthDA and failed to
be posted to AltDA, it would be automatically split and posted as
smaller batches. But automatic fallback to EthDA for arbitrarily large
batches could be very expensive and not desired, especially if the AltDA
provider was just down/unreachable temporarily or due to some
misconfiguration that could've been fixed by operators.

Now DA providers should use the ErrFallbackRequested error when
requesting fallback. Since it is sent over RPC most of the time, it is
matched with string matching inside the BatchPoster, in addition to the
error identity check.

The AnyTrust aggregator's ErrBatchToDasFailed is now wrapped with
ErrFallbackRequested by its daprovider adapter (writerForDAS).
The idea is that the AnyTrust aggregator produces metrics and
warnings/errors that operators should be able to use to detect
degredation in their DA committee before it becomes critical. If it's
really not desired to fall back automatically, operators can set
--node.batch-poster.disable-dap-fallback-store-data-on-chain and
intervene manually when batch posting fails.

This commit also updates the existing tests which expected automatic
fallback on connectivity issues to now check for failure, and adds new
tests for explicit fallback.