Clean up Aggregator constructors
The AnyTrust RPC Aggregator used to validate the batch poster's
signature on the Store request in addition to validating it on the
daserver side. In c7194300e5603e02377260cd22dca2551471a486 this was
removed but we had left the old constructors that took required L1
and sequencer inbox info, but then didn't actually use it.

This commit makes the constructor to be used outside the package
simply NewRPCAggregator, and also makes the newAggregator constructor
private.