Fix tests anytrust legacy fallback and signing (#3930)
* Disable fallback to calldata for Anytrust test

The test was PASSing even if posting batches with Anytrust was failing,
since the batch poster would fall back to calldata.

* Fix Anytrust test flakiness around fetching keyset

KeysetFetcher.GetKeysetByHash's logic for finding the SetValidKeyset
event where the keyset was created was written assuming that the
eth_getLogs From and To range was exclusive on the To side, but it's
actually inclusive. Even if tests waited for the keyset creation tx to
be included in a block, the KeysetFetcher would be trying to query the
next block too, which may not exist yet.

This change simply treats the block range as inclusive, and also adds an
additional retry loop to the test function authorizeDASKeyset that
retries until the log for the keyset creation is actually queryable.

* Fix DAS system tests default retention

The retention was set to 0 which meant that no requests were being
accepted.

* Fix Anytrust DisableSignatureChecking=true

DisableSignatureChecking=true is used in some tests and for manual
testing.

Previously on the DAS server side the SignatureVerifier was being
created with a nil seqInboxCaller in the case where signature checking
was disabled. This caused "no signature verification method configured"
errors. Now the SignatureVerifier is not created and the RPC server uses
the NoopPayloadVerifier for the streaming protocol when signature
checking is disabled.

On the client side we add a check that there is a signer if the
streaming protocol is to be used, since the signer uses the signatures
for replay protection.

* Disable altda->calldata fallback by default in tests

* Set retention in TestDASComplexConfigAndRestMirror

---------

Co-authored-by: Tsahi Zidenberg <65945052+tsahee@users.noreply.github.com>