Remove internal daprovider server for AnyTrust (#3989)
NOTE: This change applies 0195abd0eb7cedfdd915144d158667de9ae604bb,
which is part of https://github.com/OffchainLabs/nitro/pull/3949 and
was already applied to v3.8.x in https://github.com/OffchainLabs/nitro/pull/3956
but not master yet as I was hoping we could just merge #3949.
Due to trying to keep the diff as small as possible for the arbos 50
release it was not possible, so we should just merge this to master
and then backport it to v3.9.0-rc.2.

The internal daprovider server was originally added in PR #2533 to unify
the code paths between internal AnyTrust and external DA providers by
making both use RPC clients. After working with it for a while it's
clear that the drawbacks outweigh any benefit from it: it added an
unnecessary HTTP/RPC layer for in-process communication, introduced
timeout configuration complexity, and made retry behavior difficult to
reason about.

The proper abstraction point is the daprovider.Writer and Reader
interfaces, not the transport layer. External DA providers need RPC
because they're remote, but internal AnyTrust components can communicate
directly. This change removes the internal server entirely and connects
the Nitro node directly to the AnyTrust aggregator writer and REST
aggregator reader.