Only sign important fields in feed (#3996)
* Only sign important fields in feed

When signing the message in the sequencer feed, only takes in account
the relevant fields. So, if we add more fields in the future, they
shouldn't affect the message signature.

Also, replace the signature field with `signatureV2`, so old nodes won't
even know the message was signed.

Close NIT-4083

* Fix broadcast message example

* Refactor feed signature

Create a function that explicitly select which fields from the broadcast
feed message should be signed. Then, use this function when signing and
when verifying the signature.

* Add missing field: kind

* Rename signature preimage to signature hash

---------

Co-authored-by: Tsahi Zidenberg <65945052+tsahee@users.noreply.github.com>