Remove bold/util/backend.go (#4134)
* Remove bold/util/backend.go

The only thing this wrapper was doing was providing a closure of the
desiredBlockNum so that HeaderU64() doesn't require an argument.

But, that method was only called from a single end-to-end test to make it a
little more convienent to know when the parent chain had caught up with the
expected head state.

By removing the wrapper, the code is simplified and easier to maintain.

* Remove auto-added package aliases

In a follow-up PR, these packages will all be refactored such that their paths
match their package names. But, that's part of NIT-4137

* Update the customda challenge test

This also removes the import of the bold/util package which is being removed as
part of this PR.