Fix deadlock in TransactionStreamer when updating SyncMonitor
AddBroadcastMessages, ReorgAtAndEndBatch, and writeMessages all hold
insertionMutex when calling updateSyncMonitor(). However, updateSyncMonitor()
called FeedPendingMessageCount() which attempts to acquire insertionMutex
again. Since Go mutexes are not reentrant, this caused a deadlock.

The fix introduces updateSyncMonitorLocked() for use when the mutex is
already held, which calls feedPendingMessageCountImpl() instead of the
locking FeedPendingMessageCount().